requirements.txt
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
python-telegram-bot==21.7
boto3==1.35.70
requests==2.32.3
